Initial sync of thermal shield electronics
Initial sync of thermal shield electronics schematic in Altium designer.
Design is based around an auxiliary ADC (TI ADC128D818) which would
connect to the Spark core via an I2C interface.
The ADC features both differential and single ended inputs.
Differential inputs are utilized for thermocouple measurements, single
ended for thermistor measurements.
Thermistor divider resistors are TBD as common thermistors for our
application are not yet known.
The ADC should be mounted as closely to the thermocouple inputs as
possible, and the internal temperature sensor used for cold junction
compensation calculations.
Alternatively, an external thermistor could be connected to one of the
spare ADC inputs for cold junction compensation.