Initial sync of all analog solution
This electronics variant is an all analog solution which provides
filtered analog amplifiers for thermocouple inputs, and utilizes the
built-in ADC on the Spark core itself.
Also synced some documentation on thermocouple sensing.
Purposefully selected some multi-purpose I/O on the spark core for ADC
inputs such that the SPI interface is left intact.
Components not yet selected.